# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: Pa$$w0rd123
        minLength: 12
        type: string
    required:
    - email
    - password
    type: object
...
  requests.RefreshRequest:
    properties:
      token:
        example: refresh_token
        type: string
    required:
    - token
    type: object
...
  responses.Error:
    description: Error response structure with optional validation details
    properties:
      details:
        items:
          $ref: '#/definitions/responses.ValidationError'
        type: array
      message:
        example: Validation failed
        type: string
    type: object
  responses.LoginResponse:
    description: Authentication token response containing access token, refresh token
      and expiration
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      exp:
        example: 1714637422
        type: integer
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  responses.LoginResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/responses.LoginResponse'
      message:
        example: Login successful
        type: string
    type: object
...
  responses.ValidationError:
    description: Specific validation error for a single field
    properties:
      field:
        example: username
        type: string
      message:
        example: Username is required
        type: string
      tag:
        example: required
        type: string
      value:
        example: invalid_value
        type: string
    type: object
...
info:
  contact: {}
  description: This is a API Server.
  title: DockMaster Web API
  version: 0.0.1
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Perform user login
      operationId: user-login
      parameters:
      - description: User's credentials
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with login data
          schema:
            $ref: '#/definitions/responses.LoginResponseWrapper'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.Error'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/responses.Error'
        "403":
          description: Account locked
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Authenticate a user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Perform refresh access token
      operationId: user-refresh
      parameters:
      - description: Refresh token
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with new tokens
          schema:
            $ref: '#/definitions/responses.LoginResponseWrapper'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.Error'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/responses.Error'
        "403":
          description: Account locked
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Refresh access token
      tags:
      - Authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"